docker-requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
boto3==1.16.41
botocore==1.19.41
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-inherits-error-messages==1.0.0
djangorestframework==3.12.2
drf-nested-routers==0.92.5
drf-spectacular==0.13.2
drf-yasg==1.20.0
flex==6.14.1
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonpointer==2.0
MarkupSafe==1.1.1
mysqlclient==2.0.3
gevent==20.9.0
gunicorn==20.0.4
packaging==20.7
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pycryptodome==3.9.9
PyMySQL==1.0.2
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rfc3987==1.3.8
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
strict-rfc3339==0.7
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
validate-email==1.3
python-magic==0.4.18